requirements.txt
openai>=1.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
google-search-results>=2.4.0
pytest>=8.0.0
